modify contribute modal
