+++ components/Village/Styles.module.scss
@@ -182,6 +182,12 @@
       color: var(--text-color-3);
     }
 
+    h1 {
+      font-size: var(--font-header-3);
+      font-weight: 700;
+      line-height: 100%;
+    }
+
     .select {
       margin-bottom: 2rem;
     }
@@ -284,6 +290,13 @@
           }
         }
       }
+    }
+
+    ////// DETAILS INFO ///////
+    .details_info {
+      margin-top: 2rem;
+      font-size: var(--font-p-text-1);
+      font-weight: 400;
     }
 
     ////// DETAILS FORM ///////
